requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
python-dotenv